Project reorganized as an eclipse project.
Source code will be in the src/ directory.
Audio files will be in the audio/ directory. Main.java has been changed to find the audio files in the correct directory
Compiled code will be in the bin/ directory and will not be added to the repository. When you open or run the project, Eclipse will recompile the source for you.

.classpath and .project are the eclipse project files. Please do not modify them.
.gitignore is tells git to ignore files it should ignore.
please do not change anything in .git. Thank you!!